Rename component file names according to the current convention
