+++ packages/login/src/App.tsx
@@ -21,12 +21,12 @@ import { Provider } from 'react-redux'
 import { Route, Switch } from 'react-router'
 import { ConnectedRouter } from 'connected-react-router'
 import { createGlobalStyle, ThemeProvider } from 'styled-components'
-import { ForgottenItem } from './views/resetCredentialsForm/forgottenItemForm'
-import { ResetCredentialsSuccessPage } from './views/resetCredentialsForm/resetCredentialsSuccessPage'
-import { PhoneNumberVerification } from './views/resetCredentialsForm/phoneNumberVerificationForm'
-import { RecoveryCodeEntry } from './views/resetCredentialsForm/recoveryCodeEntryForm'
-import { SecurityQuestion } from './views/resetCredentialsForm/securityQuestionForm'
-import { UpdatePassword } from './views/resetCredentialsForm/updatePasswordForm'
+import { ForgottenItem } from './views/ResetCredentialsForm/ForgottenItemForm'
+import { ResetCredentialsSuccessPage } from './views/ResetCredentialsForm/ResetCredentialsSuccessPage'
+import { PhoneNumberVerification } from './views/ResetCredentialsForm/PhoneNumberVerificationForm'
+import { RecoveryCodeEntry } from './views/ResetCredentialsForm/RecoveryCodeEntryForm'
+import { SecurityQuestion } from './views/ResetCredentialsForm/SecurityQuestionForm'
+import { UpdatePassword } from './views/ResetCredentialsForm/UpdatePasswordForm'
 import { Page } from './Page'
 import { LoginBackgroundWrapper } from '@login/common/LoginBackgroundWrapper'
 import { StepTwoContainer } from '@login/views/StepTwo/StepTwoContainer'
