+++ packages/login/src/views/ResetCredentialsForm/ResetCredentialsSuccessPage.tsx
@@ -17,7 +17,7 @@ import { Box } from '@opencrvs/components/lib/Box'
 import { Stack } from '@opencrvs/components/lib/Stack'
 import { Text } from '@opencrvs/components/lib/Text'
 import { CountryLogo } from '@opencrvs/components/lib/icons'
-import { LogoContainer } from '@login/views/resetCredentialsForm/commons'
+import { LogoContainer } from '@login/views/ResetCredentialsForm/Commons'
 import React, { useEffect } from 'react'
 import { injectIntl, WrappedComponentProps as IntlShapeProps } from 'react-intl'
 import { connect } from 'react-redux'
